Rework Dockerfile
The binary is now stripped after compiling. With this extra step the
final image is just 1.88 MB in size.
Also adjusted the non root user. When creating it in stage 1 we do not
need a home dir. Furthermore we don't copy the /etc/group so there
should not be a group specified when copying the binary.